added benchmark case WIDTH=4 and moved IdentityHasher
