+++ merkle-tree/benches/hybrid_vs_plain_identity_hasher.rs
@@ -9,27 +9,28 @@ use p3_merkle_tree::{
     HybridMerkleTree, MerkleTree, SimpleHybridCompressor, UnsafeNodeConverter256BabyBearBytes,
 };
 use p3_poseidon2::{Poseidon2, Poseidon2ExternalMatrixGeneral};
-use p3_symmetric::{CompressionFunctionFromHasher, PaddingFreeSponge, TruncatedPermutation};
+use p3_symmetric::{CompressionFunctionFromHasher, IdentityHasher, TruncatedPermutation};
 use rand::{thread_rng, Rng};
 
 type BabyBearPacking = <BabyBear as Field>::Packing;
-
 type PermPoseidon =
     Poseidon2<BabyBear, Poseidon2ExternalMatrixGeneral, DiffusionMatrixBabyBear, 16, 7>;
-type HPoseidon = PaddingFreeSponge<PermPoseidon, 16, 8, 8>;
+type HIdentityBabyBear256 = IdentityHasher<8>;
 type CPoseidon = TruncatedPermutation<PermPoseidon, 2, 8, 16>;
 
 type CBlake3 = CompressionFunctionFromHasher<Blake3, 2, 32>;
 
-// Takeaway: HybridMerkleTree with unsafe compressor is faster than MerkleTree
-// with safe compressor and WIDTH=1 (but slower than WIDTH=4).
+// Takeaway: HybridMerkleTree with unsafe compressor is much faster than
+// MerkleTree with safe compressor and WIDTH=1.
 
 // COMMAND
-// sudo cargo +nightly flamegraph --bench hybrid_vs_plain_width_1 --features=unsafe-conversion
+// sudo cargo +nightly flamegraph --bench hybrid_vs_plain_width_1_identity_hasher --features=unsafe-conversion
 
 const MAX_ROWS: usize = 1 << 15;
 const MAX_COLS: usize = 2;
-const NUM_MATRICES: usize = 50;
+const NUM_MATRICES: usize = 8;
+
+// TODO allow for more matrices, control only that no more than 8 are concatenated
 
 fn get_random_leaves() -> Vec<RowMajorMatrix<BabyBear>> {
     let mut pow2_to_size = HashMap::new();
@@ -58,7 +59,7 @@ fn main() {
         &mut thread_rng(),
     );
 
-    let h_poseidon = HPoseidon::new(perm_poseidon.clone());
+    let h_identity = HIdentityBabyBear256 {};
 
     let c_poseidon = CPoseidon::new(perm_poseidon);
     let c_blake3 = CBlake3::new(Blake3 {});
@@ -71,16 +72,16 @@ fn main() {
 
     let leaves = get_random_leaves();
 
-    let mut group = criterion.benchmark_group("MerkleTree vs HybridMerkleTree");
+    let mut group = criterion.benchmark_group("MerkleTree vs HybridMerkleTree with IdentityHasher");
     group.sample_size(10);
 
     group.bench_function(
-        BenchmarkId::from_parameter("Plain compressor WIDTH = 1"),
+        BenchmarkId::from_parameter("Poseidon2 compressor, WIDTH = 1"),
         |b| {
             b.iter_batched(
                 || leaves.clone(),
                 |input| {
-                    MerkleTree::new::<BabyBear, BabyBear, _, _>(&h_poseidon, &c_poseidon, input)
+                    MerkleTree::new::<BabyBear, BabyBear, _, _>(&h_identity, &c_poseidon, input)
                 },
                 BatchSize::SmallInput,
             )
@@ -88,13 +89,13 @@ fn main() {
     );
 
     group.bench_function(
-        BenchmarkId::from_parameter("Plain compressor WIDTH = 4"),
+        BenchmarkId::from_parameter("Poseidon2 compressor, WIDTH = 4"),
         |b| {
             b.iter_batched(
                 || leaves.clone(),
                 |input| {
                     MerkleTree::new::<BabyBearPacking, BabyBearPacking, _, _>(
-                        &h_poseidon,
+                        &h_identity,
                         &c_poseidon,
                         input,
                     )
@@ -104,11 +105,14 @@ fn main() {
         },
     );
 
-    group.bench_function(BenchmarkId::from_parameter("Hybrid compressor"), |b| {
-        b.iter_batched(
-            || leaves.clone(),
-            |input| HybridMerkleTree::new(&h_poseidon, &c_hybrid_poseidon_leaves, input),
-            BatchSize::SmallInput,
-        )
-    });
+    group.bench_function(
+        BenchmarkId::from_parameter("Hybrid Blake3/{Poseidon2, WIDTH = 1} compressor"),
+        |b| {
+            b.iter_batched(
+                || leaves.clone(),
+                |input| HybridMerkleTree::new(&h_identity, &c_hybrid_poseidon_leaves, input),
+                BatchSize::SmallInput,
+            )
+        },
+    );
 }
